Adding fixleds, fixpulse, open-url-clipboard, and reset-arduino.
fixleds - there is a bug in the iwl3945 driver that causes the wifi LED
to blink, which is annoying. This script echos some stuff to /proc to
turn the LED off completely.

fixpulse - switching to pulseaudio in Intrepid means that other users on
my machine don't have permission to write to the socket PA uses. This
script fixes the permissions and groups.

open-url-clipboard - takes a url in the clipboard, removes the newline,
and runs gnome-open on it. This is a workaround for (what I feel is) a
bug in gnome-terminal.

reset-arduino - toggles the DTR bit on the arduino USB connection. This
is to work around a bug in the arduino cli environment where you can't
upload more than once per power cycle. Depends on the Device::SerialPort
Perl module.